+++ example/lib/main.dart
@@ -2,6 +2,7 @@ import 'dart:async';
 import 'dart:collection';
 
 import 'package:dengage_flutter/dengage_flutter.dart';
+import 'package:dengage_flutter/InAppInline.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 
@@ -26,7 +27,7 @@ class MyApp extends StatelessWidget {
 }
 
 class MyHomePage extends StatefulWidget {
-  MyHomePage({Key key}) : super(key: key);
+  //MyHomePage({Key key}) : super(key: key);
 
   @override
   _MyHomePageState createState() => new _MyHomePageState();
@@ -37,16 +38,18 @@ class _MyHomePageState extends State<MyHomePage> {
   String lastPush = '';
   var contactKeyController = TextEditingController();
   var lastPushController = TextEditingController();
+  HashMap<String, String> hashMap = new HashMap();
 
-static const EventChannel eventChannel = EventChannel("com.dengage.flutter/onNotificationClicked");
+  static const EventChannel eventChannel = EventChannel("com.dengage.flutter/onNotificationClicked");
+  static const EventChannel eventChannel2 = EventChannel("com.dengage.flutter/inAppLinkRetrieval");
 
-  void _onEvent(Object event) {
-    print("in on Event object is: ");
+  void _onEvent(dynamic event) {
+    print("in on Event object is: $event");
     print(event);
     lastPushController.text = event.toString();
   }
 
-  void _onError(Object error) {
+  void _onError(dynamic error) {
     print("in on Error Object is: ");
     print(error);
   }
@@ -55,15 +58,29 @@ static const EventChannel eventChannel = EventChannel("com.dengage.flutter/onNot
   void initState() {
     DengageFlutter.getContactKey().then((value) {
       print("dengageContactKey: $value");
-      contactKeyChanged(value);
+      //contactKeyChanged(value);
     });
-    DengageFlutter.startGeofence();
+    //DengageFlutter.startGeofence();
     // print("setting screen name.");
     // DengageFlutter.setNavigationWithName('MainScreen');
+    DengageFlutter.requestLocationPermissions();
+    DengageFlutter.handleNotificationActionBlock().then((value) {
+      print("dengageContactKey: $value");
+      //contactKeyChanged(value);
+    });
 
-    eventChannel.receiveBroadcastStream().listen(_onEvent, onError: _onError);
 
+    eventChannel2.receiveBroadcastStream().listen(_onEvent, onError: _onError);
+    hashMap["priya"]="priya";
     super.initState();
+  }
+
+  void showAlert(BuildContext context,String value) {
+    showDialog(
+        context: context,
+        builder: (context) => AlertDialog(
+          content: Text("hi"),
+        ));
   }
 
   contactKeyChanged(String value) {
@@ -136,10 +153,10 @@ static const EventChannel eventChannel = EventChannel("com.dengage.flutter/onNot
               padding: EdgeInsets.only(top: 10.0),
               child: ElevatedButton(
                 onPressed: () async {
-                  String msg = await DengageFlutter.getContactKey();
+                  await DengageFlutter.setLogStatus(false);
                   ScaffoldMessenger.of(context).showSnackBar(
                     SnackBar(
-                      content: Text(msg),
+                      content: Text("msg".toString()),
                       action: SnackBarAction(
                         label: 'Ok',
                         onPressed: () {},
@@ -154,15 +171,17 @@ static const EventChannel eventChannel = EventChannel("com.dengage.flutter/onNot
               padding: EdgeInsets.only(top: 10.0),
               child: ElevatedButton(
                 onPressed: () async {
-                  String token = await DengageFlutter.getToken();
+                  print("inbox");
+                  List< dynamic> s=  await DengageFlutter.getInboxMessages(0, 10);
+                  print(s.toString());
                   ScaffoldMessenger.of(context).showSnackBar(
                     SnackBar(
-                      content: Text('token: $token'),
+                      content: Text('token: $s'),
                       action: SnackBarAction(
                         label: 'copy token',
                         onPressed: () {
-                          Clipboard.setData(new ClipboardData(text: token));
-                          },
+                          Clipboard.setData(new ClipboardData(text: s.toString()));
+                        },
                       ),
                     ),
                   );
@@ -176,43 +195,34 @@ static const EventChannel eventChannel = EventChannel("com.dengage.flutter/onNot
                 onPressed: () async {
                   Map data = new HashMap<String, dynamic>();
                   data["name"] = "Kamran Younis";
-                  await DengageFlutter.sendDeviceEvent("tableName", data);
+                  await DengageFlutter.setPartnerDeviceId("hasnainHaider");
                 },
                 child: Text('Send Device Event'),
               ),
             ),
-            Container(
-              padding: EdgeInsets.only(top: 10.0),
-              child: ElevatedButton(
-                onPressed: () async {
-                  String subscription = await DengageFlutter.getSubscription();
-                  ScaffoldMessenger.of(context).showSnackBar(
-                    SnackBar(
-                      content: Text(subscription),
-                      action: SnackBarAction(
-                        label: 'copy',
-                        onPressed: () {
-                          Clipboard.setData(
-                              new ClipboardData(text: subscription));
-                        },
-                      ),
-                    ),
-                  );
-                },
-                child: Text('Get & Show subscription'),
-              ),
+            SizedBox(
+              height: 100,
+              child: Container(
+                padding: EdgeInsets.only(top: 10.0),
+                child: InAppInline(
+                  propertyId: "2233",
+                  screenName: "test",
+                  hideIfNotFound: false,
+                  customParams: hashMap,
+
+                ),),
             ),
             Container(
               padding: EdgeInsets.only(top: 10.0),
               child: ElevatedButton(
                 onPressed: () async {
-                    DengageFlutter.setTags([{
-                          "tagName": "kamranTesting",
-                          "tagValue": "My Test Tag Value",
-    // let changeTime: String?
-    // let removeTime: String?
-    // let changeValue: String?
-                    }]);
+                  DengageFlutter.setTags([{
+                    "tagName": "kamranTesting",
+                    "tagValue": "My Test Tag Value",
+                    // let changeTime: String?
+                    // let removeTime: String?
+                    // let changeValue: String?
+                  }]);
                 },
                 child: Text('set Tags'),
               ),
@@ -221,7 +231,7 @@ static const EventChannel eventChannel = EventChannel("com.dengage.flutter/onNot
               padding: EdgeInsets.only(top: 10.0),
               child: ElevatedButton(
                 onPressed: () async {
-                    Navigator.push(context, MaterialPageRoute(builder: (context) => SecondRoute()));
+                  Navigator.push(context, MaterialPageRoute(builder: (context) => SecondRoute()));
                 },
                 child: Text('Navigate to "SecondScreen"'),
               ),
@@ -250,7 +260,6 @@ static const EventChannel eventChannel = EventChannel("com.dengage.flutter/onNot
     );
   }
 }
-
 class SecondRoute extends StatelessWidget {
   SecondRoute () {
     print("setting screen name as SecondScreen");